server/ml/requirements.txt
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
imbalanced-learn==0.12.4
xgboost==2.1.4
lightgbm==4.5.0
optuna==4.1.0
shap==0.46.0
Pillow==11.0.0
tensorflow==2.21.0
fastapi==0.115.0
uvicorn==0.30.6
